Improve tag parsing. Still untested.
We now save the last tag if it ends at the end of the chunk. If it does, that might mean it will continue in the next chunk, unless it turns it there is no next chunk (EOF).

We also separate tags by `type:`, although we only use the `meta:` type for now.

I'm not sure I like this whole system, but it's the best thing I can come up with.