Reject empty entries and entries that are just whitespace.
Whitespace rejection only applies client-side, and we should take the MIME type into consideration. For example, it should be possible to use with Whitespace programming language without any problems, but you'll have to upload the files using the form, instead of composing them in the editor and submitting them as Markdown.