Rewrote repo session auth code using fibers for no reason.
The plan was to add database connection pooling at auth time. It didn't work out because pulls use sessions too, many of them for long periods of time. We can't use one DB connection per session.

But the code is simpler now, so that's a win. Right now I'm planning on dumping fibers once ECMAScript 6 is turned on by default, which should be pretty easy with one of these generator libraries. The code is almost identical to Futures.